% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{adjust.brightness}
\alias{adjust.brightness}
\title{Adjust color brightness}
\usage{
adjust.brightness(colors, d)
}
\arguments{
\item{colors}{Character vector of one or more colors in hex format}

\item{d}{Adjustment for brightness}
}
\description{
Adjust brightness of one or more colors via conversion to HSV colorspace
}
\details{
Final brightness will be bounded within [0, 1]
}
