% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{adjust.color.hsb}
\alias{adjust.color.hsb}
\title{Adjust color properties}
\usage{
adjust.color.hsb(colors, h = 0, s = 0, b = 0)
}
\arguments{
\item{colors}{Character vector of one or more colors in hex format}

\item{h}{Adjustment for hue}

\item{s}{Adjustment for saturation}

\item{b}{Adjustment for brightness}
}
\description{
Adjust hue, saturation, or brightness of one or more colors via conversion
to HSV colorspace and back
}
\details{
Final hue, saturation, and brightness will be bounded within [0, 1]
}
