% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{boxcox}
\alias{boxcox}
\title{Box-cox transformation}
\usage{
boxcox(x)
}
\arguments{
\item{x}{Vector of numeric values to transform}
}
\value{
Numeric vector of transformed values in the same order as \code{x}
}
\description{
Applies a box-cox transformation to a numeric vector
}
\seealso{
\code{\link[car:powerTransform]{car::powerTransform()}}
}
