% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{calc.hwe.xy}
\alias{calc.hwe.xy}
\title{Convert genotype frequencies to Hardy-Weinberg coordinates}
\usage{
calc.hwe.xy(f.het, f.hom)
}
\arguments{
\item{f.het}{Frequency of heterozygous genotypes}

\item{f.hom}{Frequency of homozygous alternate genotypes}
}
\value{
Numeric vector of (x, y) coordinates
}
\description{
Convert a pair of heterozgous and homozygous alt genotype frequencies to ternary
(x, y) coordinates for Hardy-Weinberg visualization
}
