% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{calc.plot.quadrant.density}
\alias{calc.plot.quadrant.density}
\title{Order plot quadrants by density}
\usage{
calc.plot.quadrant.density(x, y, decreasing = FALSE)
}
\arguments{
\item{x}{Numeric vector of scatterplot x values}

\item{y}{Numeric vector of scatterplot y values}

\item{decreasing}{Order quadrants by most to least dense [default: FALSE]}
}
\value{
Character vector of ordered plot quadrant names
}
\description{
Order Euclidean quadrants in a scatterplot based on the number of points in each
}
\details{
Uses standard Euclidean notation for quadrant names
\(e.g., "I" = top right; "IV" = bottom right\)
}
