% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{categorical.rainbow}
\alias{categorical.rainbow}
\title{Categorical rainbow palette}
\usage{
categorical.rainbow(
  n,
  hue.range = c(0, 1),
  saturation.range = c(0.5, 0.8),
  value.range = c(0.3, 0.95),
  period = 1
)
}
\arguments{
\item{n}{Number of colors in palette}

\item{hue.range}{Two-element numeric vector specifying range of eligible hues [default: c(0, 1)]}

\item{saturation.range}{Two-element numeric vector specifying range of
eligible saturations  [default: c(0, 1)]}

\item{value.range}{Two-element numeric vector specifying range of
eligible values  [default: c(0, 1)]}

\item{period}{Scalar for how many complete sine rotations (2*pi) should be sampled
for saturation and value [default: 1]}
}
\value{
Character vector of hex colors
}
\description{
Generate a rainbow color palette suitable for N categorical variables
}
\details{
Colors will be uniformly sampled across \code{hue.range}, with saturation
and value being sampled approximately according to \code{sin(x)} and \code{-sin(x)},
for x uniformly spaced elements in [0, \code{period}*pi], respectively.
}
\seealso{
\code{\link[=hsv.palette]{hsv.palette()}}
}
