% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{chisq.power.from.or}
\alias{chisq.power.from.or}
\title{Chi-square power}
\usage{
chisq.power.from.or(N, P, OR)
}
\arguments{
\item{N}{Total observations}

\item{P}{Expected proportion of observations with affirmative outcome
(e.g., disease|event) under a null baseline}

\item{OR}{Target odds ratio}

\item{alpha}{Desired type I error rate [default: 0.05]}
}
\description{
Calculate the power of a 2x2 Chi-square test given a known sample size and prespecified effect odds ratio
}
