% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean.numeric.labels}
\alias{clean.numeric.labels}
\title{Clean labels for numeric values}
\usage{
clean.numeric.labels(
  vals,
  suffix.delim = "",
  acceptable.decimals = 1,
  min.label.length = 1,
  max.label.length = Inf,
  respect.original.vals = FALSE,
  return.rounded.vals = FALSE
)
}
\arguments{
\item{vals}{Vector of numeric values to translate to labels}

\item{suffix.delim}{Character delimiter (i.e., separator) between value and
suffix [default: no delimiter, ""]}

\item{acceptable.decimals}{Number of significant decimal digits to permit
before jumping to a greater log suffix [default: 1]}

\item{min.label.length}{Minimum total count of digits to include in each
label; labels shorter than this length will have their decimal places expanded
[default: 1]}

\item{max.label.length}{Maximum total count of digits to include in each
label; labels shorter than this length will have their decimal places truncated
[default: Infinite]}

\item{respect.original.vals}{Should the exact length of \code{vals} be respected,
with no values being collapsed when they round to the same significant digit?
[default: FALSE, that is: values that round to the same significant digit
will be deduplicated]}

\item{return.rounded.vals}{Should the exact rounded numeric values also
be returned? [default: FALSE]}
}
\description{
Convert large numeric values to simple, human-readable labels with log suffixes
}
\details{
Note that all labels will be converted to the same scale / suffix scheme

If element-wise conversion is desired, it's recommended to apply over \code{vals} instead
}
\examples{

vals <- 10^(2:5)

# Standard invocation:
clean.numeric.labels(vals, acceptable.decimals=1)

# Element-wise implementation
sapply(vals, clean.numeric.labels)

}
