% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{density.topomap}
\alias{density.topomap}
\title{Density-based topographic heatmap}
\usage{
\method{density}{topomap}(
  x,
  y,
  xlims = NULL,
  ylims = NULL,
  res = 200,
  palette.fxn = viridis::mako,
  ncol = 96,
  add = TRUE,
  contours = TRUE,
  contour.col = "white",
  contour.lwd = 0.5,
  contour.levels = NULL,
  retry.with.jitter = TRUE
)
}
\arguments{
\item{x}{Point values on X axis}

\item{y}{Point values on Y axis; must match X}

\item{xlims}{Custom range of values to evaluate on X axis}

\item{ylims}{Custom range of values to evaluate on Y axis}

\item{res}{Resolution of (X, Y) grid}

\item{ncol}{Number of colors generated using \code{palette.fxn()}}

\item{add}{Should the heatmap be added to the existing plot device? [default: TRUE]}

\item{contours}{Should step contours be superimposed? [default: TRUE]}

\item{contour.col}{Color of contour lines [default: "white"]}

\item{contour.lwd}{Width of contour lines [default: 0.5]}

\item{contour.levels}{Value of \code{nlevels} passed to \code{\link[graphics:contour]{graphics::contour()}}}

\item{retry.with.jitter}{Fault tolerance (see details) [default: TRUE]}

\item{palete.fxn}{Function to generate colors (see \code{Details})}
}
\description{
Transform (x, y) value pairs into a 2D topographic heatmap, with optional contour lines
}
\details{
\code{palette.fxn} must accept a single integer value as input and return
a vector of colors of length equal to integer input (see \code{\link[viridis:reexports]{viridis::viridis()}})

When \code{x} and/or \code{y} have extremely low (near-zero) variance, \code{\link[MASS:kde2d]{MASS::kde2d()}}
can fail to estimate density due to bandwidths of zero. When \code{retry.with.jitter}
is \code{TRUE}, density estimation will be automatically retried with \code{x} and \code{y}
values artificially noised by \code{\link[=jitter]{jitter()}} with default parameters. This seems
to usually solve the \code{\link[MASS:kde2d]{MASS::kde2d()}} issue and results in a reasonably close
facimile of the true density.
}
\seealso{
\code{\link[MASS:kde2d]{MASS::kde2d()}}, \code{\link[graphics:image]{graphics::image()}}, \code{\link[graphics:contour]{graphics::contour()}}
}
