% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{density.w.outliers}
\alias{density.w.outliers}
\title{Density with Outlier Rug}
\usage{
\method{density}{w.outliers}(
  vals,
  style = "density",
  min.complexity = 30,
  bw.adj = 1,
  outlier.lower.bound = NULL,
  outlier.upper.bound = NULL,
  outlier.tick.hex = 0.02,
  color = "gray80",
  title = NULL,
  x.title = NULL,
  x.label.units = NULL,
  y.title = NULL,
  outlier.lwd = 1/3,
  parmar = c(2, 2, 0.35, 0.35)
)
}
\arguments{
\item{vals}{Numeric values to plot}

\item{style}{Style of density plot; either "density" (default) for kernel
density estimator or "hist"/"histogram" for histogram}

\item{min.complexity}{Minimum number of unique values in \code{vals} before
automatically defaulting to \code{style == "hist"} [default: 30]}

\item{bw.adj}{Bandwidth adjustment for density estimation. See \code{adjust} in \code{\link[stats:density]{stats::density()}}.}

\item{outlier.lower.bound}{Threshold below which an observation is treated
as an outlier [default: Q1 - 3*IQR]}

\item{outlier.upper.bound}{Threshold below which an observation is treated
as an outlier [default: Q3 + 3*IQR]}

\item{outlier.tick.hex}{Relative height expansion of outlier ticks [default: 0.02]}

\item{color}{Color for density area [default: "gray80"]}

\item{title}{(Optional) Title for plot}

\item{x.title}{Title for X-axis [default: no title]}

\item{x.label.units}{Units for X-axis labels; passed to \code{\link[=clean.axis]{clean.axis()}}}

\item{y.title}{Title for Y-axis [default: no title]}

\item{outlier.lwd}{Line width for outlier ticks [default: 1/3]}

\item{parmar}{Margin values passed to par()}
}
\description{
Plot one-dimensional density with a marginal rug of ticks for outlier observations
}
\details{
When \code{style == "density"}, this function will attempt to approximate
the equivalent count values for the Y-axis, although these values will depend
on \code{bw.adj} and can sometimes be misleading
}
\seealso{
\code{\link[stats:density]{stats::density()}}, \code{\link[=clean.axis]{clean.axis()}}
}
