% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{dynamic.range}
\alias{dynamic.range}
\title{Dynamic range}
\usage{
dynamic.range(values)
}
\arguments{
\item{values}{Vector of numeric values}
}
\value{
numeric ratio of max / min
}
\description{
Calculate the dynamic range of a numeric vector; that is, the ratio of the
maximum to minimum non-NA values in the vector
}
