% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{greyscale.palette}
\alias{greyscale.palette}
\title{Greyscale palette}
\usage{
greyscale.palette(n)
}
\arguments{
\item{n}{Number of colors in palette}
}
\value{
Character vector of hex colors
}
\description{
Generate a sequential greyscale palette that does not end at black & white
}
