% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{hsv.palette}
\alias{hsv.palette}
\title{HSV-based palette interpolation}
\usage{
hsv.palette(hues, sats, vals)
}
\arguments{
\item{hues}{Hue values in [0, 1]}

\item{sats}{Saturation values in [0, 1]}

\item{vals}{Brightness values in [0, 1]}
}
\value{
Vector of hex color codes
}
\description{
Generate a color palette based on uniform sampling across HSV space
}
\details{
All of \code{hues}, \code{sats}, and \code{vals} must have the same lengths. If
any of these vectors is detected to be shorter than the others, its values
will be recycled as per normal R convention.
}
\examples{
set.seed(1234)
colors <- hsv.palette(hues=(0:10)/10, sats=runif(10, 0.5, 0.8), vals=runif(10, 0.4, 0.9))
plot(1:10, 1:10, pch=19, cex=3, col=colors)

}
