% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{is.inside}
\alias{is.inside}
\title{Position-interval intersection}
\usage{
is.inside(pos, interval)
}
\arguments{
\item{pos}{Query integer}

\item{interval}{Two-element numeric vector defining the bounds of the target interval}
}
\value{
Logical
}
\description{
Logical check of whether one integer is inside an interval
defined by two other integers
}
\examples{
is.inside(2, c(1, 3))
# TRUE

is.inside(5, c(1, 3))
# FALSE

}
