% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{ivw.meta}
\alias{ivw.meta}
\title{Inverse-variance weighted meta-analysis}
\usage{
ivw.meta(estimates, vars, conf = 0.95)
}
\arguments{
\item{estimates}{Numeric point estimates (e.g., log-odds ratios)}

\item{vars}{Variances for estimates}

\item{conf}{Confidence interval width [default: 0.95]}
}
\value{
Named numeric vector with pooled point estimate, pooled variance,
lower CI bound, and upper CI bound
}
\description{
Conduct a simple inverse-variance weighted meta-analysis of point estimates
}
