% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{km.curve}
\alias{km.curve}
\title{Kaplan-Meier Curves}
\usage{
km.curve(
  surv.models,
  colors,
  group.names = NULL,
  km.lwd = 3,
  ci.alpha = 0.1,
  legend = TRUE,
  legend.names = NULL,
  legend.label.spacing = 0.075,
  legend.label.cex = 1,
  title = NULL,
  y.title = "Survival Probability",
  xlims = NULL,
  x.label.line = -0.75,
  x.title.line = 0,
  x.tck = -0.0175,
  time.is.days = TRUE,
  parmar = c(2, 3, 0.25, 4)
)
}
\arguments{
\item{surv.models}{List of one or more \code{\link[survival:summary.survfit]{survival::summary.survfit}} objects}

\item{colors}{Vector of colors for the list elements in \code{surv.models}}

\item{group.names}{(Optional) group names to assign to each list element
in \code{surv.models}}

\item{km.lwd}{Line width for Kaplan-Meier curves [default: 3]}

\item{ci.alpha}{Transparency value \code{alpha} for confidence interval
shading [default: 0.1]}

\item{legend}{Should a legend be plotted?}

\item{legend.names}{(Optional) mapping of \code{values} to labels for legend}

\item{legend.label.spacing}{Minimum vertical spacing between legend
labels [default: 0.075]}

\item{legend.label.cex}{Character expansion value for text labels in
legend [default: 1]}

\item{title}{(Optional) Title for plot}

\item{y.title}{Title for Y-axis [default: "Survival Probability"]}

\item{xlims}{(Optional) two-element vector of start and stop values for
X-axis, specified in days (or years if \code{time.is.days} is \code{FALSE})}

\item{x.label.line}{Line for X-axis labels [default: -0.75]}

\item{x.title.line}{Line for X-axis title [default: 0]}

\item{x.tck}{X-axis tick length [default: -0.0175]}

\item{time.is.days}{Should time be interpreted as days [defualt: TRUE]}

\item{parmar}{Margin values passed to par()}
}
\description{
Plot Kaplan-Meier curves for one or more datasets or strata
}
\seealso{
\code{\link[survival:Surv]{survival::Surv}}, \code{\link[survival:survfit]{survival::survfit}}, \code{\link[survival:summary.survfit]{survival::summary.survfit}}
}
