% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{label.outliers}
\alias{label.outliers}
\title{Define Outlier Values}
\usage{
label.outliers(values, n.iqr = 3, fixed.min = NULL, fixed.max = NULL)
}
\arguments{
\item{values}{Vector of numeric values}

\item{n.iqr}{Multiple of IQRs beyond which a point will be treated
as an outlier [default: 3]}

\item{fixed.min}{Optional. If specified, will override the upper bound below
which observations will be treated as outliers.}
}
\value{
Logical vector of outlier assignments for each element in \code{values}
}
\description{
Define outlier observations from a numeric vector using the standard
Q1/Q3 + N*IQR approach
}
