% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{optimize.label.color}
\alias{optimize.label.color}
\title{Optimize label color}
\usage{
optimize.label.color(bg.color, cutoff = 0.7)
}
\arguments{
\item{bg.color}{Hex code for background color upon which the label will be printed}

\item{cutoff}{Brightness threshold for switching from white to black [default: 0.7]}
}
\value{
Either "black" or "white"
}
\description{
Automatically determine whether a label should be white or black based on the
brightness of the background color upon which the label will be added
}
