% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{parse.gt}
\alias{parse.gt}
\title{Parse a VCF GT}
\usage{
parse.gt(gt, count.ref = FALSE, any.na = TRUE)
}
\arguments{
\item{gt}{A character vector of one or more VCF-style genotype string to be parsed}

\item{count.ref}{Should the number of reference alleles be returned?
[default: return count of non-ref alleles]}

\item{any.na}{Should \code{NA} be returned if any missing allele is encountered?
[default: TRUE]}
}
\value{
Integer of allele dosage
}
\description{
Converts a VCF-style GT string to integer non-ref (or, alternatively, ref) dosage
}
\examples{
parse.gt(c("0/1", "1/1", "3|4", "./4", "./.", "0|0"))

}
