% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{relu}
\alias{relu}
\title{ReLU activation function}
\usage{
relu(x)
}
\arguments{
\item{x}{Numeric values to be processed}
}
\description{
Applies the Rectified Linear Unit (ReLU) activation function to a numeric input vector
}
