% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remap}
\alias{remap}
\title{Keyed vector remapping}
\usage{
remap(x, map, default.value = NULL)
}
\arguments{
\item{x}{Vector of values (keys) to be remapped}

\item{map}{Named vector mapping keys (vector names) to values}

\item{default.value}{Optional default value to assign for elements of \code{x} not
present in \code{names(map)}. By default, values of \code{x} that fail to remap into
\code{map} will be left unchanged.}
}
\description{
Partial remapping of vectors that can tolerate data missingness
}
\details{
Inspired by Pandas map() function:
https://pandas.pydata.org/docs/reference/api/pandas.Series.map.html
}
