% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{scatterplot}
\alias{scatterplot}
\title{Two-dimensional scatterplot}
\usage{
scatterplot(
  X,
  Y,
  colors = NULL,
  title = NULL,
  x.label.line = NULL,
  x.title = NULL,
  x.title.line = 0.5,
  x.axis.units = NULL,
  xlims = NULL,
  y.label.line = NULL,
  y.title = NULL,
  y.title.line = 0.5,
  y.axis.units = NULL,
  ylims = NULL,
  legend.vals = NULL,
  legend.labels = NULL,
  cex = 0.3,
  add = FALSE,
  parmar = c(2.5, 2.5, 1, 1)
)
}
\arguments{
\item{X}{Numeric vector of values to be plotted on the X-axis (see \code{Details})}

\item{Y}{Numeric vector of values to be plotted on the Y-axis (see \code{Details})}

\item{colors}{Vector of colors. Must be supplied in the same order as \code{X} and \code{Y} [default: "gray70"]}

\item{title}{Main title for plot [default: NULL]}

\item{x.label.line}{Line for X-axis labels (\code{label.line} parameter for \link{clean.axis})}

\item{x.title}{Title for X-axis}

\item{x.title.line}{Line for X-axis title (\code{title.line} parameter for \link{clean.axis})}

\item{x.axis.units}{Value of \code{label.units} passed to \link{clean.axis} for X axis labels}

\item{xlims}{Limits for X-axis}

\item{y.label.line}{Line for X-axis labels (\code{label.line} parameter for \link{clean.axis})}

\item{y.title}{Title for Y-axis}

\item{y.title.line}{Line for axis titles (\code{title.line} parameter for \link{clean.axis})}

\item{y.axis.units}{Value of \code{label.units} passed to \link{clean.axis} for Y axis labels}

\item{ylims}{Limits for Y-axis}

\item{legend.vals}{Named vector mapping category names to colors [default: NULL]}

\item{legend.labels}{Optional vector to overwrite names of \code{legend.vals}}

\item{cex}{Character expansion factor for individual points [default: 0.3]}

\item{add}{Should this scatterplot be added on top of an existing graphics
device? [default: FALSE]}

\item{parmar}{Numeric vector of values to pass to \code{par(mar)}}
}
\description{
Generate a scatterplot of two numeric values with options for coloring
}
\details{
If \code{legend.vals} is not provided (or is \code{NULL}), no legend will be added.
}
