% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{simulate.gts}
\alias{simulate.gts}
\title{Simulate genotypes}
\usage{
\method{simulate}{gts}(af, n, seed = NULL, return.dosage = FALSE)
}
\arguments{
\item{af}{Minor allele frequency}

\item{n}{Number of genotypes to simulate}

\item{seed}{Optional random seed}

\item{return.dosage}{Return integer allele dosages [default: FALSE]}
}
\value{
Either a character vector of simulated VCF-style genotypes or a
numeric vector of allele dosages, depending on \code{return.dosage}
}
\description{
Simulate a vector of biallelic genotypes according to Hardy-Weinberg Equilibrium
}
