% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{smart.spacing}
\alias{smart.spacing}
\title{Smart Element Spacing}
\usage{
smart.spacing(
  ideal.values,
  min.dist,
  lower.limit = -Inf,
  upper.limit = Inf,
  e = 1e-07,
  max.iter = 1000
)
}
\arguments{
\item{ideal.values}{Numeric values to be spaced}

\item{min.dist}{Minimum distance between values}

\item{lower.limit}{Values are not allowed to be placed below this limit
[default: no limits]}

\item{upper.limit}{Values are not allowed to be placed above this limit
[default: no limits]}

\item{e}{Nuisance parameter used to break ties [default: 10e-8]}

\item{max.iter}{Maximum number of iterations for optimization [default: 1,000]}
}
\value{
Numeric vector
}
\description{
Enforce minimum distance between elements in a vector
}
