% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{stacked.barplot}
\alias{stacked.barplot}
\title{Stacked barplot}
\usage{
stacked.barplot(
  major.values,
  minor.values,
  minor.colors = NULL,
  x.title = NULL,
  y.label.cex = 5/6,
  bar.hex = 0.8,
  add.legend = TRUE,
  parmar = c(0.5, 3, 2.5, 0.5)
)
}
\arguments{
\item{major.values}{"Major" axis values; used for Y-axis groupings}

\item{minor.values}{"Minor" axis values; will be depicted as stacked bars
within each major axis group}

\item{minor.colors}{Color assignments for "minor" axis values; see \code{Details}}

\item{x.title}{Optional title for top X axis}

\item{y.label.cex}{Cex parameter for Y-axis "major" group labels}

\item{bar.hex}{Width of bars relative to size of gap between bars. Setting
\code{bar.hex == 1} will leave no gap between the bars}

\item{add.legend}{Should a legend of minor value colors be added to the
bottom-right of the plot? [default: add legend]}

\item{parmar}{Margin values passed to par()}

\item{details}{By default, \code{minor.colors} will uniformly sample a grayscale palette and
assign one color to each unique value present in \code{minor.values}. Alternatively,
custom color assignments can be specified in the following ways:
\enumerate{
\item As a function to generate a palette, which will be used in place of the
grayscale palette described above
\item As a named vector of color codes recognized by R, where the names of the
vector elements map onto all unique values of \code{minor.values}
}}
}
\description{
Produce a stacked barplot of two categorical variables
}
\examples{
set.seed(2024)
x <- sample(1:5, 100, replace=TRUE)
y <- sample(letters[1:4], 100, replace=TRUE)
stacked.barplot(x, y)

}
