% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sumstat.welch.test}
\alias{sumstat.welch.test}
\title{Welch's t-test for summary statistics}
\usage{
sumstat.welch.test(means, sds, ns)
}
\arguments{
\item{means}{Two-element numeric vector of sample means}

\item{sds}{Two-element numeric vector of sample standard deviations}

\item{ns}{Two-element integer vector of sample sizes}
}
\value{
data.frame with summary of test, including difference in means,
standard error, t statistic, and P-value
}
\description{
Implementation of Welch's two-sample t-test for unequal variances based on
summary statistics rather than observation-level data
}
\details{
Based on an implementation on StackExchange:
https://stats.stackexchange.com/questions/30394/how-to-perform-two-sample-t-tests-in-r-by-inputting-sample-statistics-rather-tha
}
