% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{title.case}
\alias{title.case}
\title{Title case conversion}
\usage{
title.case(x, case = "title")
}
\arguments{
\item{x}{Character vector to be converted}

\item{case}{Either \code{title} or \code{sentence}}
}
\value{
Character vector \code{x} in desired case
}
\description{
Convert a character vector to title (or, optionally, sentence) case
}
