% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{train.elastic.net.cv}
\alias{train.elastic.net.cv}
\title{Train an elastic net regression}
\usage{
train.elastic.net.cv(
  X,
  Y,
  k = 5,
  seed = 2025,
  fold.indexes = NULL,
  inner.k = k,
  tune.length = 25
)
}
\arguments{
\item{X}{Matrix of features}

\item{Y}{Vector of numeric targets}

\item{k}{Number of cross-validation folds [default: 5]}

\item{seed}{Random seed to use for cross-validation fold assignment [default: 2025]}

\item{fold.indexes}{Optional list of numeric indexes corresponding to
cross-validation folds. If provided, will override \code{k} and \code{seed}.
[default: use random CV]}

\item{inner.k}{How many inner cross-validations should be performed within
each fold? [default: same as \code{k}]}

\item{tune.length}{Resolution of grid search to optimize elastic net params [default: 25]}
}
\value{
A fitted \link[caret:train]{caret::train} list object
}
\description{
Optimize an elastic net regression with k-fold nested cross-validation
}
\seealso{
\link[caret:trainControl]{caret::trainControl}
}
