% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{greyscale.palette}
\alias{greyscale.palette}
\title{Greyscale palette}
\usage{
greyscale.palette(n, oscillate = FALSE)
}
\arguments{
\item{n}{Number of colors in palette}

\item{oscillate}{Should colors be interleaved to maximize contrast?
[default: \code{FALSE}, which provides a sequential greyscale palette]}
}
\value{
Character vector of hex colors
}
\description{
Generate a sequential greyscale palette that does not end at black & white
}
