% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{shorten.text}
\alias{shorten.text}
\title{Abbreviate text by width}
\usage{
shorten.text(text, width = Inf, cex = 1, font = 1)
}
\arguments{
\item{text}{Input text to be abbreviated}

\item{width}{Maximum width for abbreviated label, in user units}

\item{cex}{Value of \code{cex} for \code{text}}

\item{font}{Value of \code{font} for \code{text}}
}
\value{
Abbreviated version of \code{text} that can be
}
\description{
Abbreviate any string to constrain it within a prespecified width for plotting
}
\seealso{
\code{\link[graphics:strwidth]{graphics::strwidth()}}
}
