% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{stacked.barplot}
\alias{stacked.barplot}
\title{Stacked barplot}
\usage{
stacked.barplot(
  major.values,
  minor.values = NULL,
  colors = NULL,
  inner.borders = NULL,
  outer.borders = NULL,
  outer.border.lwd = 1,
  as.proportion = FALSE,
  add.major.labels = TRUE,
  x.axis.side = 3,
  x.title = NULL,
  x.title.line = 0.3,
  x.label.line = -0.65,
  x.axis.tck = -0.025,
  y.label.cex = 5/6,
  bar.hex = 0.8,
  add.legend = TRUE,
  legend.xadj = -0.075,
  major.legend = FALSE,
  major.legend.colors = NULL,
  major.legend.xadj = -0.04,
  minor.labels.on.bars = FALSE,
  minor.label.letter.width = 0.05,
  minor.label.color = NULL,
  minor.label.cex = 5/6,
  annotate.counts = FALSE,
  end.label.xadj = -0.025,
  end.label.cex = 5/6,
  orient = "right",
  custom.major.order = NULL,
  custom.minor.order = NULL,
  sort.minor = FALSE,
  parmar = c(0.5, 3, 2.5, 0.5)
)
}
\arguments{
\item{major.values}{"Major" axis values; used for Y-axis groupings}

\item{minor.values}{"Minor" axis values; if provided, will be depicted as
stacked bars within each major axis group}

\item{colors}{Color assignments for "minor" axis values; see \code{Details}}

\item{inner.borders}{Color assignments for minor segment inner.borders
[default: "white"]}

\item{outer.borders}{Color assignment for outer border [default: "black]}

\item{outer.border.lwd}{\code{lwd} parameter for outer border [default: 1]}

\item{as.proportion}{Should values be plotted on a relative (proportional /
percentage) scale? [default: FALSE, i.e., plot raw counts]}

\item{add.major.labels}{Should major (y-axis) groups be labeled in the margin?
[default: TRUE]}

\item{x.axis.side}{\code{side} value for x-axis; \code{NA} will disable X-axis plotting.
[default: 3 (top axis)]}

\item{x.title}{Optional title for X axis}

\item{x.title.line}{Value of \code{title.line} passed to \code{\link[=clean.axis]{clean.axis()}}}

\item{x.label.line}{Value of \code{label.line} passed to \code{\link[=clean.axis]{clean.axis()}}}

\item{x.axis.tck}{Value of \code{tck} passed to \code{\link[=clean.axis]{clean.axis()}}}

\item{y.label.cex}{Cex parameter for Y-axis "major" group labels}

\item{bar.hex}{Width of bars relative to size of gap between bars. Setting
\code{bar.hex == 1} will leave no gap between the bars}

\item{add.legend}{Should a legend of minor value colors be added to the
bottom-right of the plot? [default: add legend]}

\item{legend.xadj}{Legend x-position adjustment, in relative user units.
Only relevant if \code{add.legend} is \code{TRUE}. [default: -0.075]}

\item{major.legend}{Should a major color legend be added between Y axis and
labels? [default: FALSE]}

\item{major.legend.colors}{Named vector mapping \code{major.values} to colors for
\code{major.legend} [default: uniform greyscale]}

\item{major.legend.xadj}{X adjustment scalar for major legend [default: -0.04]}

\item{minor.labels.on.bars}{Should minor value labels be printed on bars,
where space is permitting? [default: FALSE]}

\item{minor.label.letter.width}{Proportion of total X-axis to apportion per
letter for minor labels. Only used if \code{minor.labels.on.bars} is \code{TRUE}. [default: 0.05]}

\item{minor.label.color}{Color for minor label text. If \code{NULL}, color will be
dynamically optimized to be most visible against each bar's background color.}

\item{minor.label.cex}{Character expansion parameter for minor label text [default: 5/6]}

\item{annotate.counts}{Should exact bar counts be annotated at the tip of
each bar? [default: no annotations]}

\item{end.label.xadj}{End-label x-position adjustment, in relative user units.
Only relevant if \code{annotate.counts} is \code{TRUE}. [default: -0.025]}

\item{end.label.cex}{Character expansion parameter for end label text [default: 5/6]}

\item{orient}{Should the bar length be increasing to the \code{right} or
\code{left}? [default: \code{right}]}

\item{custom.major.order}{Specific order of major values to use for bars}

\item{custom.minor.order}{Specific order of minor values to use for bars}

\item{sort.minor}{Should minor labels be sorted by abundance prior to plotting?
Only used if \code{custom.minor.order} is \code{NULL}. [default: alphabetically order
minor labels]}

\item{parmar}{Margin values passed to par()}

\item{details}{By default, \code{colors} will uniformly sample a grayscale palette and
assign one color to each unique value present in \code{minor.values} (or \code{major.values},
if \code{minor.values} is not specified). Alternatively, custom color assignments
can be specified in the following ways:
\enumerate{
\item As a function to generate a palette, which will be used in place of the
grayscale palette described above
\item As a named vector of color codes recognized by R, where the names of the
vector elements map onto all unique values of \code{minor.values}
}}
}
\description{
Produce a stacked barplot of one or two categorical variables
}
\examples{
set.seed(2024)
x <- sample(1:5, 100, replace=TRUE)
y <- sample(letters[1:4], 100, replace=TRUE)
stacked.barplot(x, y)

}
