% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{staple.bracket}
\alias{staple.bracket}
\title{Staple-style bracket}
\usage{
staple.bracket(x0, x1, y0, y1, orient = "vertical", staple.len = 0.25, ...)
}
\arguments{
\item{x0}{First x position for bracket}

\item{x1}{Second x position for bracket}

\item{y0}{First y position for bracket}

\item{y1}{Second y position for bracket}

\item{orient}{Orientation of bracket; either \code{"vertical"} or \code{"horizontal"} [default: vertical]}

\item{staple.len}{Length of bracket end markers [default: 0.25]}

\item{...}{Other parameters passed to \code{\link[=segments]{segments()}}}
}
\description{
Add a staple-style square bracket to a plot
}
