% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{step.function}
\alias{step.function}
\title{Step function}
\usage{
step.function(x, y, offset = 0.5, interpolate = FALSE)
}
\arguments{
\item{x}{Original x values}

\item{y}{Original y values}

\item{offset}{Scalar ~ [0, 1] that indicates relatively how far the "step"
should be extended to the right of each point. [default: 0.5]}

\item{interpolate}{Should missing points be interpolated by extending the
flanking steps to cover the missing values? [default: FALSE]}
}
\description{
Transform an (x, y) pair into a step function
}
